requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
dialogflow==0.7.2
discord.py==1.3.1
google-api-core==1.16.0
google-auth==1.11.0
googleapis-common-protos==1.51.0
grpcio==1.27.1
idna==2.8
idna-ssl==1.1.0
multidict==4.7.4
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.11.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
websockets==8.1
yarl==1.4.2
